requirements.txt
asn1crypto==1.5.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.16.1
Flask==3.0.3
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
plotly==5.24.1
py4j==0.10.9.7
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
six==1.17.0
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
tenacity==9.0.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.21.0
